GradebookNewTest/BookTest.cs
using Gradebook2;
using Xunit;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;


namespace GradeBook2.tests
{
    public class BookTest
    {
        [Fact]
        public static void TestStats()
        {
            // arrange
            var book = new Book("");
            book.AddGrade(89.1);
            book.AddGrade(90.5);
            book.AddGrade(77.3);

            // act
            var result = book.GetStatistics();

            // assert
            Assert.Equal(85.6, result.Average, 1);
            Assert.Equal(90.5, result.High, 1);
            Assert.Equal(77.3, result.Low, 1);
            Assert.Equal('B', result.Letter);
        }
    }

    [AttributeUsage(AttributeTargets.All)]
    internal class FactAttribute : Attribute
   {
    }
}
